Add a new HtspFileInputStream class
This looks and feels like normal InputStream, but behind the scenes fetches
the contents of the file over HTSP. Valid file paths are:

* /dvrfile/ID
* /imagecache/ID

This is currently a really basic implementation, we should improve this with
readahead (rather than blocking every time we consume our buffer) and by
overriding more of the read() methods, so we don't always read one byte at a
time from the buffer.

In order to implement this, we've added support for "synchronous" requests
over the HtspMessageDispatcher, where the caller will be blocked until a
response has been received.